sim_1 function
Adding a function to run a set of test sims in one function.

Uses a workaround in place of join, as Jupyter seems to not play nice with multiprocessing and joins.
To help ensure  the workaround to mirror join as much a spossible
made the put to output queue as late as possible.